api/pyproject.toml
[project]
name = "hrms-api"
version = "0.1.0"
description = "HRMS Platform - FastAPI Backend"
requires-python = ">=3.12"
dependencies = [
    # Core framework
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-multipart>=0.0.18",
    # Database & ORM
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "greenlet>=3.1",
    # Authentication & Security
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7.4",
    "pyotp>=2.9",
    # Background Jobs & Scheduling
    "celery[redis]>=5.4",
    "redis>=5.2",
    # Template & Document Generation
    "jinja2>=3.1",
    "weasyprint>=63",
    "openpyxl>=3.1",
    # File Storage
    "boto3>=1.35",
    "python-magic>=0.4.27",
    # HTTP & External APIs
    "httpx>=0.28",
    "anthropic>=0.42",
    # Email
    "aiosmtplib>=3.0",
    "email-validator>=2.2",
    # Utilities
    "python-dateutil>=2.9",
    "orjson>=3.10",
    "tenacity>=9.0",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "factory-boy>=3.3",
    "testcontainers[postgres]>=4.9",
    "faker>=33.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
